+++ Source/WOS/Managers/NetObjectManager.h
@@ -4,7 +4,6 @@
 
 #include "CoreMinimal.h"
 #include "Managers/ManagerBase.h"
-#include "GameActor/PlayerCharacter.h"
 #include "GameActor/NetObject.h"
 #include "NetObjectManager.generated.h"
 
@@ -20,13 +19,6 @@ public:
 	UNetObjectManager();
 	void HandleSpawnPlayer(uint64 ObjectId, FVector Position, FString Name, bool bIsMine = false);
 	void HandleNetObjectPacket(uint64 ObjectId, const Packet* RecievedPacket);
-public:
-	template<class T> requires std::is_base_of_v<NetObject, T>
-	TArray<T> FindObjectsByType()
-	{
-		TArray<T> Result;
-		return Result;
-	}
 private:
 	TMap<uint64, NetObject*> NetObjects;
 	TSubclassOf<ACharacter> PlayerClass;
