+++ Source/WOS/Network/Rpc.h
@@ -4,38 +4,19 @@
 
 #include "Packet.h"
 #include "GameActor/NetObject.h"
+#include "Util/Templates.h"
+
+#define THIS_CLASS std::remove_pointer<decltype(this)>::type
 
 #define RPC(Func) _RPC##Func()
-#define USE_RPC(Func) RpcView::Add(RPC(Func), this)
-#define RPC_FUNCTION(Class, Func) uint16 RPC(Func) { static uint16 RpcId = RpcView::Register(&Class::Func); return RpcId; }
+#define RPC_FUNCTION(Func) inline uint16 RPC(Func) { static uint16 RpcId = RpcView::Register(&THIS_CLASS::Func); return RpcId; }
+#define CallRPC(Func, Target, ...) Execute(&THIS_CLASS::Func, this, RPC(Func), Target, __VA_ARGS__);
 
 enum class RpcTarget : uint16
 {
 	All,
 	Other
 };
-
-template <int>
-struct variadic_placeholder {};
-
-namespace std {
-	template <int N>
-	struct is_placeholder<variadic_placeholder<N>>
-		: integral_constant<int, N + 1>
-	{
-	};
-}
-
-
-template <typename Ret, typename Class, typename... Args, size_t... Is, typename... Args2>
-auto bind(std::index_sequence<Is...>, Ret (Class::*fptr)(Args...), Args2&&... args) {
-	return std::bind(fptr, std::forward<Args2>(args)..., variadic_placeholder<Is>{}...);
-}
-
-template <typename Ret, typename Class, typename... Args, typename... Args2>
-auto bind(Ret (Class::*fptr)(Args...), Args2&&... args) {
-	return bind(std::make_index_sequence<sizeof...(Args) - sizeof...(Args2) + 1>{}, fptr, std::forward<Args2>(args)...);
-}
 
 template<class T, class... Args> requires std::is_base_of_v<NetObject, T>
 class Rpc : Packet
@@ -56,6 +37,7 @@ public:
 		{
 			if (Object != nullptr)
 				std::apply(bind(RpcFunc, Object), Params);
+			else Objects.Remove(Object);
 		}
 	}
 	void WriteParameter(Args... ArgsList)
@@ -71,13 +53,12 @@ public:
 public:
 	void SetTarget(RpcTarget NotifyTarget) { this->Target = NotifyTarget; }
 	void SetBuffer(std::span<char> Buffer) { Data() = std::vector(Buffer.begin(), Buffer.end()); }
-	void Add(void* Ptr) { Objects.Add(static_cast<T*>(Ptr)); }
+	void Add(void* Ptr) { Objects.AddUnique(static_cast<T*>(Ptr)); }
 private:
 	RpcTarget Target;
 	void(T::*RpcFunc)(Args...);
 	TArray<T*> Objects;
 };
-
 
 class RpcView
 {
@@ -102,19 +83,15 @@ public:
 		RpcFuncTable.Add(RpcId, Interface);
 		return RpcId;
 	}
-
+	
 	template<class T, class... Args>
-	static void Execute(uint16 Id, RpcTarget Target, Args&&... ArgsList)
+	static void Execute(void(T::*)(Args...), T* Owner, uint16 Id, RpcTarget Target, Args... ArgsList)
 	{
 		auto RpcObj = static_cast<Rpc<T, Args...>*>(RpcFuncTable[Id].Owner);
 		RpcObj->SetTarget(Target);
 		RpcObj->WriteParameter(Forward<Args>(ArgsList)...);
+		RpcObj->Add(Owner);
 		SendRpc(reinterpret_cast<Packet*>(RpcObj));
-	}
-	
-	static void Add(uint16 Id, void* Owner)
-	{
-		RpcFuncTable[Id].Add(Owner);
 	}
 
 	static void RecvRPC(std::span<char> buffer,  uint16 Id)
