+++ Source/WOS/Util/Templates.cpp
@@ -1,0 +1,1 @@
+﻿#include "Templates.h"
